+++ R/geographies.R
@@ -24,7 +24,7 @@ get_statcan_geographies <- function(census_year,level,type="cartographic",
                                     cache_path = NULL,timeout=1000,
                                     refresh=FALSE,quiet=FALSE) {
   valid_census_years <- c("2021")
-  valid_levels <- c("PR","CD","CMACA","CMA","CA","CSD","CT","ADA","DA","ER","FED","DPL","POPCNTR","POPCTR","FSA")
+  valid_levels <- c("PR","CD","CMACA","CMA","CA","CSD","CT","ADA","DA","ER","FED","DPL","POPCNTR","POPCTR","FSA","HR")
   valid_types <- c("cartographic","digital")
   if (!(census_year %in% valid_census_years)) {
     stop(paste0("Census year must be one of ",paste0(valid_census_years,collapse = ", "),"."))
@@ -51,6 +51,13 @@ get_statcan_geographies <- function(census_year,level,type="cartographic",
       stop(paste0("Problem, don't know how to get geographic data for level ",level,"."))
     }
     url <- paste0("https://www12.statcan.gc.ca/census-recensement/",census_year,"/geo/sip-pis/boundary-limites/files-fichiers/l",tolower(level),filler,typeID,"21a_e.zip")
+    if (level=="HR") {
+      if (type=="cartographic") {
+        url <- "https://www150.statcan.gc.ca/n1/en/pub/82-402-x/2024001/hrbf-flrs/carto/ArcGIS/HR_000b23a_e.zip?st=wQRRL7PQ"
+      } else {
+        url <- "https://www150.statcan.gc.ca/n1/en/pub/82-402-x/2024001/hrbf-flrs/digi/ArcGIS/HR_000a23a_e.zip"
+      }
+    }
     tmp <- tempfile()
     options(timeout = timeout)
     utils::download.file(url,tmp,mode="wb",quiet=quiet)